+++ packages/typespec-autorest/src/openapi.ts
@@ -233,6 +233,8 @@ export interface AutorestDocumentEmitterOptions {
    * @default "for-visibility-only"
    */
   readonly emitCommonTypesSchema?: "never" | "for-visibility-changes";
+
+  readonly xmlStrategy: "xml-service" | "none";
 }
 
 /**
@@ -305,6 +307,7 @@ export async function getOpenAPIForService(
   const auth = processAuth(service.type);
 
   const xml = await resolveXmlModule();
+  const xmlStrategy = options.xmlStrategy;
 
   const root: OpenAPI2Document = {
     swagger: "2.0",
@@ -370,16 +373,18 @@ export async function getOpenAPIForService(
   const routes = httpService.operations;
   reportIfNoRoutes(program, routes);
 
+  const xmlEnabled = xmlStrategy !== "none";
+
   // The set of produces/consumes values found in all operations
   let allResponseContentTypes = routes
     .flatMap((route) => route.responses)
     .flatMap((res) => res.responses)
-    .flatMap((res) => res.body?.contentTypes)
+    .flatMap((res) => res.body?.contentTypes ?? [])
     .filter(
       (ct) =>
         // XML and JSON are privileged to be the only content types that can live at the top level and inform global
         // serializer/naming behavior.
-        ct === "application/json" || ct === "application/xml",
+        ct === "application/json" || (xmlEnabled && ct === "application/xml"),
     );
   if (allResponseContentTypes.length === 0) allResponseContentTypes = ["application/json"];
   const globalProduces = new Set<string>(allResponseContentTypes);
@@ -387,12 +392,16 @@ export async function getOpenAPIForService(
   let allRequestContentTypes = routes
     .flatMap((route) => route.parameters)
     .flatMap((param) => param.body?.contentTypes ?? [])
-    .filter((ct) => ct === "application/json" || ct === "application/xml");
+    .filter(
+      (ct) => !!ct && (ct === "application/json" || (xmlEnabled && ct === "application/xml")),
+    );
   if (allRequestContentTypes.length === 0) allRequestContentTypes = ["application/json"];
   const globalConsumes = new Set<string>(allRequestContentTypes);
 
-  const specHasXml = globalProduces.has("application/xml") || globalConsumes.has("application/xml");
+  const shouldEmitXml =
+    xmlEnabled && (globalProduces.has("application/xml") || globalConsumes.has("application/xml"));
   const specIsOnlyXml =
+    xmlEnabled &&
     globalProduces.size === 1 &&
     globalProduces.has("application/xml") &&
     globalConsumes.size === 1 &&
@@ -1328,6 +1337,10 @@ export async function getOpenAPIForService(
       schema: bodySchema,
     };
 
+    if (shouldFlattenProperty(context.tcgcSdkContext, param) === true) {
+      result["x-ms-client-flatten"] = true;
+    }
+
     const jsonName = resolveEncodedName(program, param, "application/json");
     if (jsonName !== param.name) {
       // Special case to be able to keep pre-existing cases where you have both the body parameter name and x-ms-client-name
@@ -1553,7 +1566,7 @@ export async function getOpenAPIForService(
         checkDuplicateTypeName(program, processed.type, name, root.definitions!);
         processed.ref.value = "#/definitions/" + encodeURIComponent(name);
         if (processed.schema) {
-          if (specHasXml)
+          if (shouldEmitXml)
             attachXml(
               processed.type,
               name,
@@ -2010,7 +2023,7 @@ export async function getOpenAPIForService(
         property["x-ms-client-name"] = clientName;
       }
 
-      if (specHasXml && xml.available) {
+      if (shouldEmitXml && xml.available) {
         const xmlName = resolveEncodedName(program, prop, "application/xml");
 
         if (xmlName !== propertySchemaName) {
@@ -2133,7 +2146,7 @@ export async function getOpenAPIForService(
       applyArmIdentifiersDecorator(prop.type, propSchema, prop);
     }
 
-    if (specHasXml && xml.available) {
+    if (shouldEmitXml && xml.available) {
       attachPropertyXml(prop, propSchema);
     }